requirements.txt
# Python dependencies for Nightingale VoiceAI
# Healthcare voice AI with privacy-first architecture
# Updated for Python 3.11 compatibility

# Core web framework
fastapi==0.104.1
uvicorn==0.24.0

# Data validation and settings
pydantic==2.4.2
pydantic-settings==2.0.3

# Database and ORM
sqlalchemy==2.0.23
alembic==1.12.1

# Authentication and security
pyjwt==2.8.0
cryptography==41.0.7
passlib[bcrypt]==1.7.4

# Natural language processing (updated for better compatibility)
spacy>=3.7.0,<3.8.0
# Install spacy model separately with: python -m spacy download en_core_web_sm

# Speech processing (updated torch versions for better compatibility)
openai-whisper==20231117
torch>=2.0.0,<2.2.0
torchaudio>=2.0.0,<2.2.0

# HTTP client for API calls
httpx==0.25.2
aiofiles==23.2.1

# Testing framework
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# System monitoring
psutil==5.9.6

# Development and debugging
python-multipart==0.0.6
python-jose[cryptography]==3.3.0

# Logging and monitoring
structlog==23.2.0

# UI Framework and Visualization
streamlit==1.49.1
plotly==6.3.0
matplotlib>=3.10.0
seaborn>=0.13.0

# Optional: Database drivers
# sqlite3 (built into Python)
# Uncomment for PostgreSQL production deployment:
# psycopg2-binary==2.9.9

# Optional: Redis for caching
# redis==5.0.1

# Environment management
python-dotenv==1.0.0